+++ scripts/macros.js
@@ -7,6 +7,8 @@ export {auraDamageEnd} from './macros/2014/monsterFeatures/generic/auraDamage.js
 export {autoGrapple} from './macros/2014/monsterFeatures/generic/autoGrapple.js';
 export {autoPush} from './macros/2014/monsterFeatures/generic/autoPush.js';
 export {berserk} from './macros/2014/monsterFeatures/generic/berserk.js';
+export {bloodFrenzy} from './macros/2014/monsterFeatures/generic/bloodFrenzy.js';
+export {bloodiedFrenzy} from './macros/2014/monsterFeatures/generic/bloodiedFrenzy.js';
 export {contextualBonus} from './macros/2014/monsterFeatures/generic/contextualBonus.js';
 export {damageTurnStart} from './macros/2014/monsterFeatures/generic/damageTurnStart.js';
 export {deathBurst} from './macros/2014/monsterFeatures/generic/deathBurst.js';