+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3.38 Release:
+## New Generic Monster Features:
+- Blood Frenzy
+- Bloodied Frenzy
+  
 # 1.3.37 Release:
 ## New Modern Features:
 - Stroke of Luck